Move dynamic delay test to own package. Wrap checked exceptions as unchecked in applying delay function.
